+++ Cargo.toml
@@ -99,6 +99,7 @@ members = [
     "rpc-client",
     "rpc-client-api",
     "rpc-client-nonce-utils",
+    "rpc-request",
     "rpc-test",
     "runtime",
     "runtime-transaction",
@@ -563,6 +564,7 @@ solana-rpc = { path = "rpc", version = "=2.2.0" }
 solana-rpc-client = { path = "rpc-client", version = "=2.2.0", default-features = false }
 solana-rpc-client-api = { path = "rpc-client-api", version = "=2.2.0" }
 solana-rpc-client-nonce-utils = { path = "rpc-client-nonce-utils", version = "=2.2.0" }
+solana-rpc-request = { path = "rpc-request", version = "=2.2.0" }
 solana-runtime = { path = "runtime", version = "=2.2.0" }
 solana-runtime-transaction = { path = "runtime-transaction", version = "=2.2.0" }
 solana-sdk = { path = "sdk", version = "=2.2.0" }
