+++ rpc-client-api/Cargo.toml
@@ -11,8 +11,6 @@ edition = { workspace = true }
 
 [dependencies]
 anyhow = { workspace = true }
-base64 = { workspace = true }
-bs58 = { workspace = true, features = ["std"] }
 jsonrpc-core = { workspace = true }
 reqwest = { workspace = true, default-features = false, features = ["rustls-tls"] }
 reqwest-middleware = { workspace = true }
@@ -20,23 +18,16 @@ semver = { workspace = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
 serde_json = { workspace = true }
-solana-account = { workspace = true }
 solana-account-decoder-client-types = { workspace = true }
 solana-clock = { workspace = true }
-solana-commitment-config = { workspace = true, features = ["serde"] }
 solana-fee-calculator = { workspace = true, features = ["serde"] }
 solana-inflation = { workspace = true }
-solana-inline-spl = { workspace = true }
-solana-pubkey = { workspace = true }
+solana-rpc-request = { workspace = true }
 solana-signer = { workspace = true }
 solana-transaction-error = { workspace = true }
 solana-transaction-status-client-types = { workspace = true }
 solana-version = { workspace = true }
 thiserror = { workspace = true }
-
-[dev-dependencies]
-const_format = { workspace = true }
-solana-pubkey = { workspace = true, features = ["rand"] }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
