+++ rpc-client-api/src/lib.rs
@@ -1,10 +1,8 @@
 #![allow(clippy::arithmetic_side_effects)]
 
 pub mod client_error;
-pub mod config;
 pub mod custom_error;
 pub mod error_object;
-pub mod filter;
 pub mod request;
 pub mod response;
