+++ rpc-request/src/filter.rs
@@ -1,6 +1,6 @@
 use {
     base64::{prelude::BASE64_STANDARD, Engine},
-    serde::Deserialize,
+    serde_derive::{Deserialize, Serialize},
     solana_account::{AccountSharedData, ReadableAccount},
     solana_inline_spl::{token::GenericTokenAccount, token_2022::Account},
     std::borrow::Cow,
@@ -91,7 +91,7 @@ pub enum MemcmpEncodedBytes {
     Bytes(Vec<u8>),
 }
 
-impl<'de> Deserialize<'de> for MemcmpEncodedBytes {
+impl<'de> serde::Deserialize<'de> for MemcmpEncodedBytes {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: serde::Deserializer<'de>,
